requirements.txt
azure-identity==1.5.0
azure-keyvault-secrets==4.2.0
openpyxl==3.0.9
pandas==1.1.5
python-dotenv==0.15.0
requests==2.25.1

